% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anoes.R
\name{anoes.geeglm}
\alias{anoes.geeglm}
\title{Analysis of Effect Sizes (ANOES) based on the Robust Effect Size index (RESI) for GEE models
This function will estimate RESI and its CI for each factor in a fitted GEE model object.
The CIs are calculated via non-parametric bootstraps.}
\usage{
\method{anoes}{geeglm}(object, alpha = 0.05, nboot = 1000)
}
\arguments{
\item{object}{the model object}

\item{alpha}{numeric, the type I error rate based on which CIs were constructed. By default, 0.05.}

\item{nboot}{numeric, the number of bootstraps used to construct CIs. By default, 1000.}
}
\value{
An ANOVA-type model summary output with RESI estimates and CIs added.
}
\description{
Analysis of Effect Sizes (ANOES) based on the Robust Effect Size index (RESI) for GEE models
This function will estimate RESI and its CI for each factor in a fitted GEE model object.
The CIs are calculated via non-parametric bootstraps.
}
