% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anoes.glm.R
\name{anoes.glm}
\alias{anoes.glm}
\title{Analysis of Effect Sizes based on the Robust Effect Size index (RESI) for (generalized) linear regression models}
\usage{
anoes.glm(
  model.full,
  model.reduced = NULL,
  .vcov = sandwich::vcovHC,
  method = "bootstrap",
  nboot = 1000,
  alpha = 0.05
)
}
\arguments{
\item{model.full}{model formula for full model}

\item{model.reduced}{the reduced model or comparison model. By dafult, NULL}

\item{.vcov}{the covariance estimator for tested coefficient(s), by default, sandwich::vcovHC will be used.}

\item{method}{The method constructing CI.
"wild.norml" = the wild bootstrapping with standard normal multiplier;
"wild.rademacher" = wild bootstrapping with multiplier following rademacher distribution;
"f" = non-central F distribution;
"chisq" = non-cental Chi-square distribution
By default, the non-parametric bootstrap will be used.}

\item{nboot}{if bootstrapped CI was specified in `method`, nboot = the number of bootstraps that will be implemented. By default, 1000 bootstraps will be applied.}

\item{alpha}{The significance level of the constructed CI. By default, 5%}

\item{data}{the data frame containing the variables speficiedd in the model formula(s)}
}
\value{

}
\description{
Analysis of Effect Sizes based on the Robust Effect Size index (RESI) for (generalized) linear regression models
}
