% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{ARMtx}
\alias{ARMtx}
\title{Functions for simulations}
\usage{
ARMtx(time, rho.e)
}
\arguments{
\item{time}{the values of time variables for a subject}

\item{rho.e}{the correlation between two errors}
}
\description{
Function genearting longitudinal data
based on linear mixed-effects model.
}
\details{
Function for the correlation structure for the errors within a subject
}
