% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{bayes.samp}
\alias{bayes.samp}
\title{Bayesian bootstrap sampling (Rubin, 1981)}
\usage{
bayes.samp(data)
}
\arguments{
\item{data}{data frame; The data frame that need bootstrapping}
}
\value{
Returns a data frame containing weights generated via Bayesian bootstraps.
}
\description{
Bayesian bootstrap sampling (Rubin, 1981)
}
