% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.ci.R
\name{boot.ci}
\alias{boot.ci}
\title{Function generating CIs for the RESI via various bootstrap methods}
\usage{
boot.ci(
  model.full,
  model.reduced = NULL,
  r = 1000,
  robust.var = TRUE,
  multi = "none",
  boot.type = 1,
  alpha = 0.05,
  correct = FALSE,
  num.cores = 1,
  digits = 3
)
}
\arguments{
\item{model.full}{the full model. It should be a `glm` object.}

\item{model.reduced}{the reduced `glm()` model to compare with the full model. By default `NULL`, it's the same model as the full model but only having intercept.}

\item{r}{numeric, the number of boostrap replicates. By default, 1000 bootstraps will be implemented.}

\item{robust.var}{default to TRUE, whether to use the robust (sandwich) variance estimator when construct the Wald test statistic. If `TRUE`, the variance of the estimator will be obtained by using `sandwich::vcovHC()`` and the HC3 will be applied.}

\item{multi}{the distribution from which the multipliers will be drawn: 'none' = the multipliers equal constant 1 (default); 'rad' = rademacher; 'normal' = Std Normal distribution}

\item{boot.type}{which type of bootstrap to use. 1: resampling covariates along with residuals (default); 2: fixing covariates and only bootstrapping residulas; 3: resampling covariates and residuals independently w/ replacements; 4. no sampling, just multipliers}

\item{alpha}{significance level of the constructed CIs. By default, 0.05 will be used0}

\item{correct}{for the linear regression models (i.e., `family = 'gaussian'` in the `glm()` function) whether the residuals with bias correction will be used, by default, FALSE.}

\item{num.cores}{The number of CPU cores to be used for calculating bootstrapped CIs, by default, only 1 core will be used.}

\item{digits}{the number of decimal digits in the output ANOES table. By default, 3}

\item{sigma2}{the true variance of error under homoskedasticity assumption (added for simulations).}
}
\description{
Function generating CIs for the RESI via various bootstrap methods
}
