% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{boot.samp}
\alias{boot.samp}
\title{Non-parametric bootstrap sampling}
\usage{
boot.samp(data, id.var = NULL)
}
\arguments{
\item{data}{data frame; The data frame that need bootstrapping}

\item{id.var}{character; for clustered/longitudinal data, the name of id variable used as sampling unit.}
}
\value{
Returns a data frame containing bootstrapped data.
}
\description{
Non-parametric bootstrap sampling
}
