% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_resi.R
\name{calc_resi.lm}
\alias{calc_resi.lm}
\title{Robust effect size add-on for lm and glm
Effect size summary that can compliment your summary output. All conversions are based on the robust effect size index estimator.}
\usage{
\method{calc_resi}{lm}(object, object.reduced = NULL, vcov. = sandwich::vcovHC, ...)
}
\arguments{
\item{object}{The model object.}

\item{object.reduced}{The reduced model to compare with the object while calculating the effect size. By default, `model.reduced = NULL`; if a reduced model is specified, only the effect size for the different parameters will be shown.}

\item{vcov.}{The variance covariance matrix to use. Defaults to one of the heteroskedasticity consistent estimator. Note that this differs from the default in lmtest.}

\item{...}{Arguments passed to `vcov.`}
}
\value{
Returns a summary of tests and the robust effect size index for linear models.
}
\description{
Robust effect size add-on for lm and glm
Effect size summary that can compliment your summary output. All conversions are based on the robust effect size index estimator.
}
\details{
Stuff and stuff.
}
\keyword{effect}
\keyword{lm}
\keyword{size}
