% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_resi.R
\name{calc_resi.lme}
\alias{calc_resi.lme}
\title{RESI from lme object
This function calculate the RESI from lme model object}
\usage{
\method{calc_resi}{lme}(object, ...)
}
\arguments{
\item{object}{The lme model object}
}
\value{
returns the ANOVA-type summary table with RESI estimate for each factor
}
\description{
RESI from lme object
This function calculate the RESI from lme model object
}
