% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_resi.R
\name{calc_resi.lmerMod}
\alias{calc_resi.lmerMod}
\title{RESI for lmer object}
\usage{
\method{calc_resi}{lmerMod}(object, robust.var = TRUE, ...)
}
\arguments{
\item{object}{The lmer model object}
}
\value{
returns the ANOVA-type summary table with RESI estimate for each factor
}
\description{
RESI for lmer object
}
