% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R, R/S.R
\name{chisq2S}
\alias{chisq2S}
\title{This function computes the robust effect size index from Vandekar, Rao, & Blume (2020).
Vector arguments are accepted. If different length arguments are passed they are dealt with in the usual way of R.
For mixed effects models, RESI is conditional on the average correlation
structure within subjects.}
\usage{
chisq2S(chisq, df, n)

chisq2S(chisq, df, n)
}
\arguments{
\item{chisq}{The chi-square statistic for the parameter of interest.}

\item{df}{Number of degrees of freedom of the chi-square statistic.}

\item{n}{Number of independent samples.}
}
\value{
Returns a scalar or vector argument of the the robust effect size index estimate.

Returns a scalar or vector argument of the the robust effect size index estimate.
}
\description{
This function computes the robust effect size index from Vandekar, Rao, & Blume (2020).
Vector arguments are accepted. If different length arguments are passed they are dealt with in the usual way of R.
For mixed effects models, RESI is conditional on the average correlation
structure within subjects.
}
\keyword{power}
