% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{depression}
\alias{depression}
\title{Depression Treatment Data}
\format{
A data frame with 1020 rows and 5 variables:
\describe{
\item{diagnose}{diagnosed depression severity}
\item{drug}{treatment; standard or new}
\item{id}{patient id}
\item{time}{time point of treatment}
\item{depression}{depression response at time of treatment. 1 = Normal, 0 = Abnormal}
}
}
\source{
\url{http://static.lib.virginia.edu/statlab/materials/data/depression.csv}
}
\usage{
depression
}
\description{
A longitudinal dataset comparing two treatments for depression.
}
\references{
Agresti, A. (2002). Categorical Data Analysis. Wiley, 2nd Edition.
}
\keyword{datasets}
