% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ess.R
\name{ess.geeglm}
\alias{ess.geeglm}
\title{Calculating the effective sample size from a fitted GEE model.}
\usage{
\method{ess}{geeglm}(obj, xTab, robust.var = TRUE)
}
\arguments{
\item{obj}{a fitted `geeglm` object}

\item{constr}{matrix, linear constraint}
}
\description{
Calculating the effective sample size from a fitted GEE model.
}
