% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ess.R
\name{ess.lme}
\alias{ess.lme}
\title{This function calculates the effective sample size from a fitted linear mixed effect model object (lme) or a GEE model object (gee)}
\usage{
\method{ess}{lme}(obj, total = TRUE, ...)
}
\arguments{
\item{obj}{`lme` or `gee` object}

\item{constr}{matrix. The linear constraint used to form an hypothesis regarding which the RESI is built. For example, for parameters `beta`, `constr = diag(c(1, 1, 1))` means testing `constr %*% beta = 0`
By default, `constr = NULL` where the ESS for each of the parameters will be computed.}
}
\description{
This function calculates the effective sample size from a fitted linear mixed effect model object (lme) or a GEE model object (gee)
}
