% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ess.R
\name{ess.lmerMod}
\alias{ess.lmerMod}
\title{Calculating the effective sample size from a fitted LME model.}
\usage{
\method{ess}{lmerMod}(obj, xTab, robust.var = TRUE)
}
\arguments{
\item{obj}{a fitted `lmer`(`lmerMod`) object}

\item{constr}{matrix, linear constraint}
}
\description{
Calculating the effective sample size from a fitted LME model.
}
