% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nccd.R
\name{F}
\alias{F}
\title{Chi distribution function.}
\usage{
F(chisq, p, lambda)
}
\arguments{
\item{p}{integer, degrees of freedom.}

\item{lambda, }{noncentrality parameter of Chi-sq distribution.}

\item{u}{numeric, value of Chi-sq random variable.}
}
\description{
Code taken by SNV from https://www1.maths.leeds.ac.uk/~john/software/ncc/ncc.r.
}
