% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{f2S}
\alias{f2S}
\title{Compute the robust effect size index estimator from F-statistic}
\usage{
f2S(f, df, rdf)
}
\arguments{
\item{f}{The chi-square statistic for the parameter of interest.}

\item{df}{Number of degrees of freedom of the chi-square statistic.}

\item{rdf}{Model residual degrees of freedom or number of independent samples.}
}
\value{
Returns a scalar or vector argument of the the robust effect size index estimate.
}
\description{
This function computes the robust effect size index from Vandekar, Rao, & Blume (2020).
Vector arguments are accepted. If different length arguments are passed they are dealt with in the usual way of R.
}
\keyword{power}
