% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{fsq2S}
\alias{fsq2S}
\title{Covert Cohen's \emph{f}^2 to S}
\usage{
fsq2S(fsq)
}
\arguments{
\item{fsq}{Numeric, value of Cohen's \emph{f}^2.}
}
\value{
Returns an estimate the robust effect size index
}
\description{
Converts Cohen's \emph{f}^2 to robust effect size index (S)
using the formula from Vandekar, Rao, & Blume (2020).
}
\keyword{power}
