% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nccd.R
\name{gb}
\alias{gb}
\title{Function for MD Bessel interval}
\usage{
gb(u, nu, lambda)
}
\arguments{
\item{u}{something.}

\item{nu}{something else.}

\item{lambda}{parameter value.}
}
\description{
Code taken by SNV from https://www1.maths.leeds.ac.uk/~john/software/ncc/ncc.r.
}
