% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nccd.R
\name{gr}
\alias{gr}
\title{Function for MD radial interval}
\usage{
gr(u, nu, lambda)
}
\arguments{
\item{u}{something.}

\item{nu}{something else.}

\item{lambda}{parameter value.}
}
\value{
xxxx
}
\description{
Code taken by SNV from https://www1.maths.leeds.ac.uk/~john/software/ncc/ncc.r.
}
