% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{insurance}
\alias{insurance}
\title{US Health Insurance Data}
\format{
A data frame with 1338 rows and 7 variables:
\describe{
\item{age}{age of primary beneficiary in years}
\item{sex}{insurance contractor sex, male/female}
\item{bmi}{body mass index}
\item{children}{number of dependents}
\item{smoker}{smoker/non-smoker}
\item{region}{beneficiary's region of US}
\item{charges}{individual medical costs billed by health insurance}
}
}
\source{
\url{https://www.kaggle.com/datasets/teertha/ushealthinsurancedataset}
}
\usage{
insurance
}
\description{
A dataset with 1338 observations on health insurance charges and demographic factors.
}
\keyword{datasets}
