% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncf.R
\name{lambound_f}
\alias{lambound_f}
\title{Finding the bound of non-centrality parameter}
\usage{
lambound_f(y, df1, df2, alpha)
}
\arguments{
\item{y}{numeric, value of F random variable;}

\item{df1}{integer, degrees of freedom 1.}

\item{df2}{integer, degrees of freedom 2.}

\item{alpha}{significence level}
}
\description{
Finding the bound of non-centrality parameter
}
