% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nccd.R
\name{lamfind_chisq}
\alias{lamfind_chisq}
\title{Finds root of distribution function minus alpha.}
\usage{
lamfind_chisq(y, p, alpha)
}
\arguments{
\item{y}{numeric, value of chi random variable.}

\item{p}{integer, degrees of freedom.}

\item{alpha}{probability.}
}
\value{
xxxx
}
\description{
Code taken by SNV from https://www1.maths.leeds.ac.uk/~john/software/ncc/ncc.r.
}
