% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncf.R
\name{lamfind_f}
\alias{lamfind_f}
\title{Finds root of distribution function minus alpha.}
\usage{
lamfind_f(y, df1, df2, alpha)
}
\arguments{
\item{y}{numeric, value of F random variable.}

\item{df1}{integer, degrees of freedom 1.}

\item{df2}{integer, degrees of freedom 2.}

\item{alpha}{probability.}
}
\description{
Original code taken by SNV from https://www1.maths.leeds.ac.uk/~john/software/ncc/ncc.r.
}
