% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncf.R
\name{ncf.ci}
\alias{ncf.ci}
\title{Confidence interval for F noncentraliy parameter
Written by Kaidi based on Kent's code
Original kent's Code taken by SNV from https://www1.maths.leeds.ac.uk/~john/software/ncc/ncc.r.}
\usage{
ncf.ci(y, df1, df2, alpha = 0.05)
}
\arguments{
\item{y}{numeric, value of F random variable.}

\item{df1}{integer, degrees of freedom 1.}

\item{df2}{integer, degrees of freedom 2.}

\item{alpha}{two-tailed probability for confidence interval.}
}
\value{
xxx
}
\description{
Modified based on functions ncc.ci.sr() and ncc.ci.central()
}
