% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resi.R
\name{resi.geeglm}
\alias{resi.geeglm}
\title{RESI from gee or geeglm object
This function calculate the RESI from geeglm model object}
\usage{
\method{resi}{geeglm}(object, ...)
}
\arguments{
\item{object}{The model object}
}
\value{
returns the ANOVA-type summary table with RESI estimate for each factor
}
\description{
RESI from gee or geeglm object
This function calculate the RESI from geeglm model object
}
