% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resi.R
\name{resi.glm}
\alias{resi.glm}
\title{Robust effect size add-on for glm}
\usage{
\method{resi}{glm}(object, vcov. = sandwich::vcovHC, ...)
}
\arguments{
\item{object}{The model object.}

\item{vcov.}{The variance covariance matrix to use. Defaults to one of the heteroskedasticity consistent estimator. Note that this differs from the default in lmtest.}

\item{...}{Arguments passed to coeftest}
}
\value{
Returns a summary of tests and the robust effect size index for generalized linear models.
}
\description{
Effect size summary that can compliment your summary output. All conversions are based on the robust effect size index estimator.
}
\details{
Stuff and more stuff.
}
\keyword{effect}
\keyword{glm}
\keyword{size}
