% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resi.R
\name{resi.lm}
\alias{resi.lm}
\title{Analysis of Effect Sizes (ANOES) based on the Robust Effect Size index (RESI) for (generalized) linear regression models
This function estiamtes RESI and it CIs in a fitted glm model object
The CI are calculated via non-parametric bootstraps}
\usage{
\method{resi}{lm}(
  model.full,
  model.reduced = NULL,
  data,
  anova = TRUE,
  coefficients = TRUE,
  nboot = 1000,
  boot.method = "nonparam",
  vcovfunc = sandwich::vcovHC,
  alpha = 0.05,
  store.boot = FALSE,
  Anova.args = list(),
  vcov.args = list(),
  unbiased = TRUE,
  ...
)
}
\arguments{
\item{model.full}{the full model. It should be a `glm` object.}

\item{model.reduced}{the reduced `glm()` model to compare with the full model. By default `NULL`, it's the same model as the full model but only having intercept.}

\item{nboot}{numeric, the number of bootstrap replicates. By default, 1000 bootstraps will be implemented.}

\item{boot.method}{which type of bootstrap to use: `nonparam` = non-parametric bootstrap (default); `bayes` = Bayesian bootstrap.}

\item{alpha}{significance level of the constructed CIs. By default, 0.05 will be used.}

\item{robust.var}{default to TRUE, whether to use the robust (sandwich) variance estimator when construct the Wald test statistic. If `TRUE`, the variance of the estimator will be obtained by using `sandwich::vcovHC()`` and the HC3 will be applied.}
}
\description{
Analysis of Effect Sizes (ANOES) based on the Robust Effect Size index (RESI) for (generalized) linear regression models
This function estiamtes RESI and it CIs in a fitted glm model object
The CI are calculated via non-parametric bootstraps
}
