% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resi.pe.R
\name{resi.pe}
\alias{resi.pe}
\title{Robust Effect Size index (RESI) point estimation}
\usage{
resi.pe(model.full, ...)
}
\arguments{
\item{model.full}{the full model.}

\item{...}{Other arguments to be passed to Anova function}

\item{model.reduced}{the reduced model to compare with the full model. By default `NULL`, it's the same model as the full model but only having intercept.}

\item{data}{optional data frame or object coercible to data frame of model.full data (required for some model types)}

\item{anova}{whether to produce an Anova table with the RESI columns added. By default = `TRUE`}

\item{summary}{whether to produce a summary table with the RESI columns added. By default = `TRUE`}

\item{vcovfunc}{the variance estimator function for constructing the Wald test statistic. By default, sandwich::vcovHC (the robust (sandwich) variance estimator)}
}
\value{
Returns a list of RESI point estimates
}
\description{
Robust Effect Size index (RESI) point estimation
}
