% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resi.R
\name{resi}
\alias{resi}
\title{Robust Effect Size index (RESI) point and interval estimation for models}
\usage{
resi(
  model.full,
  model.reduced = NULL,
  data,
  anova = TRUE,
  summary = TRUE,
  nboot = 1000,
  vcovfunc = sandwich::vcovHC,
  alpha = 0.05,
  ...
)
}
\arguments{
\item{model.full}{the full model.}

\item{model.reduced}{the reduced model to compare with the full model. By default `NULL`, it's the same model as the full model but only having intercept.}

\item{data}{optional data frame or object coercible to data frame of model.full data (required for some model types)}

\item{anova}{whether to produce an Anova table with the RESI columns added. By default = `TRUE`}

\item{summary}{whether to produce a summary table with the RESI columns added. By default = `TRUE`}

\item{nboot}{numeric, the number of bootstrap replicates. By default, 1000 bootstraps will be implemented.}

\item{vcovfunc}{the variance estimator function for constructing the Wald test statistic. By default, sandwich::vcovHC (the robust (sandwich) variance estimator)}

\item{alpha}{significance level of the constructed CIs. By default, 0.05 will be used.}

\item{...}{Other arguments to be passed to Anova function}
}
\value{
Returns a list that includes function arguments, RESI point estimates, and confidence intervals
}
\description{
Robust Effect Size index (RESI) point and interval estimation for models
}
