% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resi.R
\name{resi.waldtest}
\alias{resi.waldtest}
\title{Robust effect size add-on for Wald tests and anova}
\usage{
\method{resi}{waldtest}(object, ..., vcov. = sandwich::vcovHC)
}
\arguments{
\item{object}{The model object.}

\item{...}{Arguments passed to waldtest}

\item{vcov.}{The variance covariance matrix to use. Defaults to one of the heteroskedasticity consistent estimator. Note that this differs from the default in lmtest.}
}
\value{
Returns an anova-like test and the robust effect size index for comparing two or more models.
}
\description{
Effect size summary that can compliment your summary output. All conversions are based on the robust effect size index estimator. Object handling is handled by the 'lmtest' package.
}
\keyword{anova}
\keyword{test,}
\keyword{wald}
