% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{Rsq2S}
\alias{Rsq2S}
\title{Covert R^2 to S}
\usage{
Rsq2S(Rsq)
}
\arguments{
\item{Rsq}{Numeric, R^2}
}
\value{
Returns an estimate of R^2 based on the RESI
}
\description{
Converts R^2, the partial coefficient of determination, to
robust effect size index (S) using the formula from Vandekar, Rao, & Blume (2020).
}
\keyword{power}
