% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S.R
\name{S2d}
\alias{S2d}
\title{Convert S to Cohen's d}
\usage{
S2d(S, pi = 0.5)
}
\arguments{
\item{S}{The robust effect size index.}

\item{pi}{The sampling proportions.}
}
\value{
Returns an estimate the robust effect size index
}
\description{
Converts the robust effect size index to Cohen's d using formula from Vandekar, Rao, & Blume (2020).
}
\details{
The pi parameter comes from the fact that Cohen's d doesn't account for unequal sample proportions in the population, but S does.
The default is set to a natural value 1/2, which corresponds to a case control design, for example, where sampling proportions always are controlled by the experimenter.
}
\keyword{power}
