% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{S2Rsq}
\alias{S2Rsq}
\title{Covert S to R^2}
\usage{
S2Rsq(S)
}
\arguments{
\item{S}{Numeric, the robust effect size index.}
}
\value{
Returns an estimate of R^2 based on the RESI
}
\description{
Converts robust effect size index (S) to R^2, the partial
coefficient of determination, using the formula from Vandekar, Rao, & Blume (2020).
}
\keyword{power}
