% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sim_data_cont}
\alias{sim_data_cont}
\title{Function for simulating longitudinal continuous outcomes}
\usage{
sim_data_cont(
  N,
  S,
  pi,
  ni_range,
  rho.G,
  sigma0,
  sigma.t,
  sigma.e,
  rho.e,
  fixed.design = TRUE
)
}
\arguments{
\item{N}{total sample size (i.e., num of subjects)}

\item{S}{the values of RESI for the fixed effects}

\item{pi}{proportion or probability of being assigned to trt group}

\item{ni_range}{range of num of measurements}

\item{rho.G}{correlation coef between random intercepts and slopes}

\item{sigma0}{SD of random intercepts}

\item{sigma.t}{SD of random slopes}

\item{sigma.e}{SD of errors}

\item{rho.e}{correlation coef of the errors within a subject}

\item{fixed.design}{whether the trt assignment is fixed by design (TRUE) or random. use togther with pi}
}
\description{
Function for simulating longitudinal continuous outcomes
}
