% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.resi.R
\name{summary.resi}
\alias{summary.resi}
\title{Summary method for resi objects}
\usage{
\method{summary}{resi}(object, alpha = NULL, ...)
}
\arguments{
\item{object}{an object resulting from resi function}

\item{alpha}{an optional new specification for the confidence level. Can be vector-valued}

\item{...}{ignored}
}
\value{
Returns a `summary.resi` object containing the computed summary table
}
\description{
After running the \code{\link{resi}} function on a fitted model, this function can be used to print the summary component. If the resi function was run with the `store.boot = TRUE` option to store the full matrix of bootstrapped estimates, the user can specify a different alpha level for this function's confidence intervals.
}
