% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{z2S}
\alias{z2S}
\title{Compute the robust effect size index estimator from Z statistic}
\usage{
z2S(z, n)
}
\arguments{
\item{z}{The Z statistic for the parameter of interest.}

\item{n}{Number of independent samples.}
}
\value{
Returns a scalar or vector argument of the the robust effect size index estimate.
}
\description{
This function computes the robust effect size index from Vandekar, Rao, & Blume (2020).
Vector arguments are accepted. If different length arguments are passed they are dealt with in the usual way of R.
}
\details{
This function computes S, the RESI, from a Z statistic using the default
formula. There is another function, \code{\link{z2S_alt}}, that uses an alternative
formula. This function's formula is derived by solving the expected value of the
Z statistic for S. It is unbiased and consistent.
}
\keyword{power}
