% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anoes.R
\name{anoes.glm}
\alias{anoes.glm}
\title{}
\usage{
\method{anoes}{glm}(
  model.full,
  model.reduced = NULL,
  robust.var = TRUE,
  boot.type = 1,
  multi = "none",
  nboot = 1000,
  alpha = 0.05
)
}
\description{

}
