% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anoes.R
\name{anoes}
\alias{anoes}
\title{Analysis of Effect Sizes (ANOES) based on the Robust Effect Size index (RESI) for (generalized) linear regression models}
\usage{
anoes(x, ...)
}
\description{
Analysis of Effect Sizes (ANOES) based on the Robust Effect Size index (RESI) for (generalized) linear regression models
}
