% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anoes.R
\name{anoes}
\alias{anoes}
\title{Analysis of Effect Sizes (ANOES) based on the Robust Effect Size index (RESI) for (generalized) linear regression models}
\usage{
anoes(
  model.full,
  model.reduced = NULL,
  data,
  anova = TRUE,
  summary = TRUE,
  nboot = 1000,
  vcovfunc = sandwich::vcovHC,
  multi = "none",
  boot.type = 1,
  alpha = 0.05,
  correct = FALSE,
  num.cores = 1,
  ...
)
}
\arguments{
\item{model.full}{the full model. It should be a `glm` object.}

\item{model.reduced}{the reduced `glm()` model to compare with the full model. By default `NULL`, it's the same model as the full model but only having intercept.}

\item{data}{optional data frame or object coercible to data frame of model.full data (required for linear models with splines)}

\item{anova}{whether to compute an ANOES table based on the Anova function. By default = `TRUE`}

\item{summary}{whether to produce a summary output with the RESI columns added. By default = `TRUE`}

\item{nboot}{numeric, the number of bootstrap replicates. By default, 1000 bootstraps will be implemented.}

\item{vcovfunc}{the variance estimator function for constructing the Wald test statistic. By default, sandwich::vcovHC (the robust (sandwich) variance estimator)}

\item{multi}{the distribution from which the multipliers will be drawn: 'none' = the multipliers equal constant 1 (default); 'rad' = rademacher; 'normal' = Std Normal distribution}

\item{boot.type}{which type of bootstrap to use. 1: resampling covariates along with residuals (default); 2: fixing covariates and only bootstrapping residulas; 3: resampling covariates and residuals independently w/ replacements; 4. no sampling, just multipliers}

\item{alpha}{significance level of the constructed CIs. By default, 0.05 will be used0}

\item{correct}{for the linear regression models (i.e., `family = 'gaussian'` in the `glm()` function) whether the residuals with bias correction will be used, by default, FALSE.}

\item{num.cores}{The number of CPU cores to be used for calculating bootstrapped CIs, by default, only 1 core will be used.}

\item{...}{Other arguments to be passed to Anova function}
}
\value{

}
\description{
Analysis of Effect Sizes (ANOES) based on the Robust Effect Size index (RESI) for (generalized) linear regression models
}
