% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anova.resi.R
\name{Anova.resi}
\alias{Anova.resi}
\title{Anova method for resi objects}
\usage{
\method{Anova}{resi}(resi.obj, alpha = NULL)
}
\arguments{
\item{resi.obj}{an object resulting from resi function}

\item{alpha}{an optional new specification for the confidence level. Can be vector-valued}
}
\value{
Returns an `anova` object containing the computed Anova-style table
}
\description{
After running the \code{\link{resi}} function on a fitted model, this function can be used to print the Anova-style table component. If the resi function was run with the `store.boot = TRUE` option to store the full matrix of bootstrapped estimates, the user can specify a different alpha level for this function's confidence intervals.
}
