% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.ci.R
\name{boot.ci}
\alias{boot.ci}
\title{function generating CIS for the RESI either via bootstraps}
\usage{
boot.ci(
  model.full,
  model.reduced = NULL,
  r = 1000,
  robust.var = TRUE,
  multi = "none",
  boot.type = 1,
  alpha = 0.05,
  correct = FALSE,
  num.cores = 1
)
}
\arguments{
\item{model.full}{the full `glm()` model}

\item{model.reduced}{the reduced `glm()` model to compare with}

\item{r}{numeric, the number of boostrap replicates. By default, 1000 bootstraps will be implemented.}

\item{robust.var}{default to TRUE, whether to use the robust (sandwich) variance estimator when construct the Wald test statistic
"F" corresponds to using var-cov estimator `vcov`, and the resulting statistic will be F-statistic;
"Chisq" corresponds to using "sandwich::vcovHC" and results a Chi-squared statistics;
"Known" corresponds to knowing the true error of the linear model with homoskedasticity. With method = "Known", sigma2 need to be specified.}

\item{multi}{the distribution from which the multipliers will be drawn: 'none' = the multipliers equal constant 1 (default); 'rad' = rademacher; 'normal' = Std Normal distribution}

\item{boot.type}{which type of bootstrap to use.}

\item{alpha}{significance level used to compute the CIs. By default, 0.05 will be used}

\item{correct}{whether the residuals with bias correction will be used, by default, TRUE.}

\item{num.cores}{The number of CPU cores to be used for calculating bootstrapped CIs, by default, only 1 core will be used.}

\item{sigma2}{the true variance of error under homoskedasticity assumption (added for simulations).}
}
\description{
function generating CIS for the RESI either via bootstraps
}
