% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_resi.R
\name{calc_resi.geeglm}
\alias{calc_resi.geeglm}
\title{Robust effect size add-on for glm}
\usage{
\method{calc_resi}{geeglm}(object, ...)
}
\arguments{
\item{object}{The model object}
}
\value{
returns the ANOVA-type summary table with RESI estimate for each factor
}
\description{
Effect size summary that can compliment your summary output. All conversions are based on the robust effect size index estimator.
RESI from gee or geeglm object
This function calculate the RESI from geeglm model object
}
