% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{chisq2Ssq}
\alias{chisq2Ssq}
\title{Transformation from Wald test statistics to squared RESI}
\usage{
chisq2Ssq(chisq, df, n)
}
\arguments{
\item{chisq}{The chi-square statistic for the parameter(s) of interest.}

\item{df}{Number of degrees of freedom of the chi-square statistic.}

\item{n}{Number of independent samples.}
}
\value{
Returns a scalar or vector argument of the squared robust effect size index estimate.
}
\description{
Transformation from Wald test statistics to squared RESI
}
