% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ess.R
\name{ess}
\alias{ess}
\alias{ess.lme}
\alias{ess.lmerMod}
\alias{ess.geeglm}
\title{Calculating Effective Sample Size}
\usage{
ess(obj, ...)

\method{ess}{lme}(obj, total = TRUE, ...)

\method{ess}{lmerMod}(obj, xTab, robust.var = TRUE, ...)

\method{ess}{geeglm}(obj, total = TRUE, ...)
}
\arguments{
\item{obj}{`lme` or `gee` object}

\item{...}{ignored}

\item{total}{whether to calculate RESI for each covariate}

\item{xTab}{xTab}

\item{robust.var}{whether robust covariance was used}
}
\description{
This function calculates the effective sample size (ESS) from Kang et. al (2022)
}
\section{Methods (by class)}{
\itemize{
\item \code{lme}: Calculates the effective sample size from a fitted linear mixed effect model object (lme) or a GEE model object (gee)

\item \code{lmerMod}: Calculates the effective sample size from a fitted LME model.

\item \code{geeglm}: Calculates the effective sample size from a fitted GEE model.
}}

