% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nccd.R
\name{Finv}
\alias{Finv}
\title{chi quantile function.}
\usage{
Finv(prob, p, lambda)
}
\arguments{
\item{prob}{numeric in [0,1], quantile of chi random variable.}

\item{p}{integer, degrees of freedom.}

\item{lambda, }{noncentrality parameter of chi distribution.}
}
\description{
Code taken by SNV from https://www1.maths.leeds.ac.uk/~john/software/ncc/ncc.r.
}
