% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resi.R
\name{resi.lm}
\alias{resi.lm}
\title{Robust effect size add-on for lm
Effect size summary that can compliment your summary output. All conversions are based on the robust effect size index estimator.}
\usage{
\method{resi}{lm}(object, vcov. = sandwich::vcovHC, ...)
}
\arguments{
\item{object}{The model object.}

\item{vcov.}{The variance covariance matrix to use. Defaults to one of the heteroskedasticity consistent estimator. Note that this differs from the default in lmtest.}

\item{...}{Arguments passed to coeftest}
}
\value{
Returns a summary of tests and the robust effect size index for linear models.
}
\description{
Robust effect size add-on for lm
Effect size summary that can compliment your summary output. All conversions are based on the robust effect size index estimator.
}
\details{
Stuff and stuff.
}
\keyword{effect}
\keyword{lm}
\keyword{size}
