% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resi.R
\name{resi}
\alias{resi}
\alias{resi.default}
\alias{resi.nls}
\alias{resi.hurdle}
\alias{resi.geeglm}
\alias{resi.lme}
\title{Robust Effect Size index (RESI) point and interval estimation for models}
\usage{
resi(model.full, ...)

\method{resi}{default}(
  model.full,
  model.reduced = NULL,
  data,
  anova = TRUE,
  summary = TRUE,
  nboot = 1000,
  boot.method = "nonparam",
  vcovfunc = sandwich::vcovHC,
  alpha = 0.05,
  ...
)

\method{resi}{nls}(
  model.full,
  model.reduced = NULL,
  data,
  summary = TRUE,
  nboot = 1000,
  boot.method = "nonparam",
  vcovfunc = sandwich::vcovHC,
  alpha = 0.05,
  ...
)

\method{resi}{hurdle}(
  model.full,
  model.reduced = NULL,
  data,
  summary = TRUE,
  nboot = 1000,
  boot.method = "nonparam",
  vcovfunc = sandwich::sandwich,
  alpha = 0.05,
  ...
)

\method{resi}{geeglm}(model.full, alpha = 0.05, nboot = 1000)

\method{resi}{lme}(model.full, alpha = 0.05, nboot = 1000)
}
\arguments{
\item{model.full}{the full model object}

\item{...}{Other arguments to be passed to Anova function}

\item{model.reduced}{the reduced model to compare with the full model. By default `NULL`, it's the same model as the full model but only having intercept.}

\item{data}{optional data frame or object coercible to data frame of model.full data (required for some model types)}

\item{anova}{whether to produce an Anova table with the RESI columns added. By default = `TRUE`}

\item{summary}{whether to produce a summary table with the RESI columns added. By default = `TRUE`}

\item{nboot}{numeric, the number of bootstrap replicates. By default, 1000.}

\item{boot.method}{which type of bootstrap to use: `nonparam` = non-parametric bootstrap (default); `bayes` = Bayesian bootstrap.}

\item{vcovfunc}{the variance estimator function for constructing the Wald test statistic. By default, sandwich::sandwich (the robust variance estimator)}

\item{alpha}{numeric, significance level of the constructed CIs. By default, 0.05.}
}
\value{
Returns a list that includes function arguments, RESI point estimates, and confidence intervals
}
\description{
Robust Effect Size index (RESI) point and interval estimation for models
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Robust Effect Size index (RESI) point and interval estimation for models

\item \code{nls}: Robust Effect Size index (RESI) point and interval estimation for nls models

\item \code{hurdle}: Robust Effect Size index (RESI) point and interval estimation for hurdle models

\item \code{geeglm}: Robust Effect Size index (RESI) point and interval estimation for GEE models
This function will estimate RESI and its CI for each factor in a fitted GEE model object.
The CIs are calculated via non-parametric bootstraps.

\item \code{lme}: Robust Effect Size index (RESI) point and interval estimation for LME models
This function will estimate RESI and its CI for each factor in a fitted LME model object.
The CIs are calculated via non-parametric bootstraps.
}}

