% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resi.R
\name{resi}
\alias{resi}
\alias{resi.default}
\alias{resi.nls}
\alias{resi.survreg}
\alias{resi.coxph}
\alias{resi.hurdle}
\alias{resi.zeroinfl}
\alias{resi.geeglm}
\alias{resi.gee}
\alias{resi.lme}
\title{Robust Effect Size Index (RESI) point and interval estimation for models}
\usage{
resi(model.full, ...)

\method{resi}{default}(
  model.full,
  model.reduced = NULL,
  data,
  anova = TRUE,
  summary = TRUE,
  nboot = 1000,
  boot.method = "nonparam",
  vcovfunc = sandwich::vcovHC,
  alpha = 0.05,
  store.boot = FALSE,
  Anova.args = list(),
  vcov.args = list(),
  z2S.form = 1,
  t2S.form = 1,
  ...
)

\method{resi}{nls}(
  model.full,
  model.reduced = NULL,
  data,
  summary = TRUE,
  nboot = 1000,
  boot.method = "nonparam",
  vcovfunc = regtools::nlshc,
  alpha = 0.05,
  store.boot = FALSE,
  vcov.args = list(),
  t2S.form = 1,
  ...
)

\method{resi}{survreg}(
  model.full,
  model.reduced = NULL,
  data,
  anova = TRUE,
  summary = TRUE,
  nboot = 1000,
  boot.method = "nonparam",
  vcovfunc = vcov,
  alpha = 0.05,
  store.boot = FALSE,
  Anova.args = list(),
  z2S.form = 1,
  ...
)

\method{resi}{coxph}(
  model.full,
  model.reduced = NULL,
  data,
  anova = TRUE,
  summary = TRUE,
  nboot = 1000,
  boot.method = "nonparam",
  vcovfunc = vcov,
  alpha = 0.05,
  store.boot = FALSE,
  Anova.args = list(),
  z2S.form = 1,
  ...
)

\method{resi}{hurdle}(
  model.full,
  model.reduced = NULL,
  data,
  summary = TRUE,
  nboot = 1000,
  boot.method = "nonparam",
  vcovfunc = sandwich::sandwich,
  alpha = 0.05,
  store.boot = FALSE,
  vcov.args = list(),
  z2S.form = 1,
  ...
)

\method{resi}{zeroinfl}(
  model.full,
  model.reduced = NULL,
  data,
  summary = TRUE,
  nboot = 1000,
  boot.method = "nonparam",
  vcovfunc = sandwich::sandwich,
  alpha = 0.05,
  store.boot = FALSE,
  vcov.args = list(),
  z2S.form = 1,
  ...
)

\method{resi}{geeglm}(model.full, alpha = 0.05, nboot = 1000, ...)

\method{resi}{gee}(model.full, data, alpha = 0.05, nboot = 1000, ...)

\method{resi}{lme}(
  model.full,
  alpha = 0.05,
  nboot = 1000,
  vcovfunc = clubSandwich::vcovCR,
  vcov.args = list(),
  ...
)
}
\arguments{
\item{model.full}{\code{lm, glm, nls, survreg, coxph, hurdle, zeroinfl, gee, geeglm} or \code{lme} model object.}

\item{...}{Ignored.}

\item{model.reduced}{Fitted model object of same type as model.full. By default `NULL`; the same model as the full model but only having intercept.}

\item{data}{Data.frame or object coercible to data.frame of model.full data (required for some model types).}

\item{anova}{Logical, whether to produce an Anova table with the RESI columns added. By default = `TRUE`.}

\item{summary}{Logical, whether to produce a summary (coefficients) table with the RESI columns added. By default = `TRUE`.}

\item{nboot}{Numeric, the number of bootstrap replicates. By default, 1000.}

\item{boot.method}{String, which type of bootstrap to use: `nonparam` = non-parametric bootstrap (default); `bayes` = Bayesian bootstrap.}

\item{vcovfunc}{The variance estimator function for constructing the Wald test statistic. By default, sandwich::vcovHC (the robust (sandwich) variance estimator).}

\item{alpha}{Numeric, significance level of the constructed CIs. By default, 0.05.}

\item{store.boot}{Logical, whether to store all the bootstrapped estimates. By default, `FALSE`.}

\item{Anova.args}{List, additional arguments to be passed to \link[car]{Anova} function.}

\item{vcov.args}{List, additional arguments to be passed to vcovfunc.}

\item{z2S.form}{Numeric, 1 (default) or 2. Formula to be used to convert Z statistic to RESI. See details.}

\item{t2S.form}{Numeric, 1 (default) or 2. Formula to be used to convert T statistic to RESI. See details.}
}
\value{
Returns a list that includes function arguments, RESI point estimates,
and confidence intervals in summary/anova-style tables
}
\description{
This function will estimate the robust effect size (RESI) from Vandekar, Rao, & Blume (2020) and its confidence interval in various ways for a fitted model object. The overall RESI is estimated via a Wald test. RESI is (optionally) estimated for each factor in summary-style table. RESI is (optionally) estimated for each variable/interaction in an Anova-style table for models with existing Anova methods. CIs can be calculated using either non-parametric or Bayesian bootstrapping.
}
\details{
The RESI, denoted as S, is applicable across many model types. It is a unitless
index and can be easily be compared across models. The RESI can also be
converted to Cohen's \emph{d} (\code{\link{S2d}}) under  model homoskedasticity.

This function computes the RESI point estimates and bootstrapped confidence
intervals based on Chi-square, F, T, or Z statistics. The robust (sandwich)
variance is used by default, allowing for consistency under
model-misspecification. The RESI is related to the non-centrality parameter
of the test statistic. The RESI estimate is consistent for all four
(Chi-square, F, T, and Z) types of statistics used. The Chi-square and F-based
calculations rely on asymptotic theory, so they may be biased in small samples.
When possible, the T and Z statistics are used. There are two formulas for both
the T and Z statistic conversion. The first, option 1 for `t2S.form` and `z2S.form`,
are based on solving the expected value of the T or Z statistic for the RESI.
The second (option 2) is based on squaring the T or Z statistic and using the
F or Chi-square statistic conversion. Both of these methods are consistent, but
option 2 exhibits a notable amount of finite sample bias, so option 1 is the
default. Option 2 may be appealing because its absolute value will be equal to
the RESI based on the F or Chi-square statistic. The RESI based on the Chi-Square
and F statistics is always greater than or equal to 0. The type of statistic
used is listed with the output. See \code{\link{f2S}}, \code{\link{chisq2S}},
\code{\link{t2S}} (option 1), \code{\link{z2S}} (option 1), \code{\link{t2S_alt}}
(option 2), and \code{\link{z2S_alt}} (option 2) for more details on the formulas.

For most model types supported by this package, two bootstrap options are
available. The default is the standard non-parametric bootstrap. Bayesian
bootstrapping is currently available for models other than \code{survreg} and
\code{coxph}.

Certain model types require the data used for the model be entered as an argument.
These are: \code{nls, survreg,} and \code{coxph}. Additionally, if a model
includes certain functions (splines, factor, I), the data needs to be provided.

If running into convergence issues with nls models, it is advised to refit the
nls model with starting values equal to the estimates provided by the model
and then try rerunning \code{resi}.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: RESI point and interval estimation for models

\item \code{nls}: RESI point and interval estimation for nls models

\item \code{survreg}: RESI point and interval estimation for survreg models

\item \code{coxph}: RESI point and interval estimation for coxph models

\item \code{hurdle}: RESI point and interval estimation for hurdle models

\item \code{zeroinfl}: RESI point and interval estimation for zeroinfl models

\item \code{geeglm}: RESI point and interval estimation for GEE models

\item \code{gee}: RESI point and interval estimation for GEE models

\item \code{lme}: RESI point and interval estimation for LME (nlme) models
}}

\examples{
# fit linear model
mod = lm(charges ~ region * age + bmi + sex, data = RESI::insurance)

# run resi on fitted model with 500 bootstrap replicates
resi(mod, nboot = 500)

# for some model types and formula structures, data argument is required
library(splines)
mod = glm(smoker ~ ns(age) + region, data = RESI::insurance, family = "binomial")

# store bootstrap results for calculating different CIs later
# specify additional arguments to the variance-covariance function via vcov.args
resi.obj = resi(mod, data = RESI::insurance, store.boot = TRUE, alpha = 0.01,
vcov.args = list(type = "HC0"))
summary(resi.obj, alpha = c(0.05, 0.1))

}
\references{
Vandekar S, Tao R, Blume J. A Robust Effect Size Index. \emph{Psychometrika}. 2020 Mar;85(1):232-246. doi: 10.1007/s11336-020-09698-2.

Kang, K., Armstrong, K., Avery, S., McHugo, M., Heckers, S., & Vandekar, S. (2021). Accurate confidence interval estimation for non-centrality parameters and effect size indices. \emph{arXiv preprint arXiv:2111.05966}.
}
\seealso{
\code{\link{resi_pe}}, \code{\link{vcovHC}}, \code{\link{nlshc}},
\code{\link{f2S}}, \code{\link{chisq2S}}, \code{\link{z2S}}, \code{\link{t2S}}
}
\concept{RESI functions}
