% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resi.R, R/resi_KK.R
\name{resi}
\alias{resi}
\alias{resi.default}
\alias{resi.nls}
\title{Robust Effect Size index (RESI) point and interval estimation for models}
\usage{
resi(x, ...)

\method{resi}{default}(
  model.full,
  model.reduced = NULL,
  data,
  anova = TRUE,
  summary = TRUE,
  nboot = 1000,
  boot.method = "nonparam",
  vcovfunc = sandwich::vcovHC,
  alpha = 0.05,
  ...
)

\method{resi}{nls}(
  model.full,
  model.reduced = NULL,
  data,
  summary = TRUE,
  nboot = 1000,
  boot.method = "nonparam",
  vcovfunc = sandwich::vcovHC,
  alpha = 0.05,
  ...
)

resi(x, ...)
}
\arguments{
\item{...}{Other arguments to be passed to Anova function}

\item{model.full}{the full model.}

\item{model.reduced}{the reduced model to compare with the full model. By default `NULL`, it's the same model as the full model but only having intercept.}

\item{data}{optional data frame or object coercible to data frame of model.full data (required for some model types)}

\item{anova}{whether to produce an Anova table with the RESI columns added. By default = `TRUE`}

\item{summary}{whether to produce a summary table with the RESI columns added. By default = `TRUE`}

\item{nboot}{numeric, the number of bootstrap replicates. By default, 1000 bootstraps will be implemented.}

\item{boot.method}{which type of bootstrap to use: `nonparam` = non-parametric bootstrap (default); `bayes` = Bayesian bootstrap.}

\item{vcovfunc}{the variance estimator function for constructing the Wald test statistic. By default, sandwich::vcovHC (the robust (sandwich) variance estimator)}

\item{alpha}{significance level of the constructed CIs. By default, 0.05 will be used.}
}
\value{
Returns a list that includes function arguments, RESI point estimates, and confidence intervals
}
\description{
Robust Effect Size index (RESI) point and interval estimation for models

Analysis of Effect Sizes (ANOES) based on the Robust Effect Size index (RESI) for (generalized) linear regression models
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Robust Effect Size index (RESI) point and interval estimation for models

\item \code{nls}: Robust Effect Size index (RESI) point and interval estimation for nls models
}}

