% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resi.R
\name{resi}
\alias{resi}
\alias{resi.default}
\alias{resi.nls}
\alias{resi.zeroinfl}
\alias{resi.geeglm}
\alias{resi.lme}
\title{Robust Effect Size index (RESI) point and interval estimation for models}
\usage{
resi(model.full, ...)

\method{resi}{default}(
  model.full,
  model.reduced = NULL,
  data,
  anova = TRUE,
  summary = TRUE,
  nboot = 1000,
  boot.method = "nonparam",
  vcovfunc = sandwich::vcovHC,
  alpha = 0.05,
  ...
)

\method{resi}{nls}(
  model.full,
  model.reduced = NULL,
  data,
  summary = TRUE,
  nboot = 1000,
  boot.method = "nonparam",
  vcovfunc = sandwich::vcovHC,
  alpha = 0.05,
  ...
)

\method{resi}{zeroinfl}(
  model.full,
  model.reduced = NULL,
  data,
  summary = TRUE,
  nboot = 1000,
  boot.method = "nonparam",
  vcovfunc = sandwich::sandwich,
  alpha = 0.05,
  ...
)

\method{resi}{geeglm}(model.full, alpha = 0.05, nboot = 1000)

\method{resi}{lme}(model.full, alpha = 0.05, nboot = 1000)
}
\arguments{
\item{model.full}{\code{lm, glm, nls, survreg, coxph, hurdle, zeroinfl, gee, geeglm} or \code{lme} model object.}

\item{...}{Other arguments to be passed to Anova function.}

\item{model.reduced}{Fitted model object of same type as model.full. By default `NULL`; the same model as the full model but only having intercept.}

\item{data}{Data.frame or object coercible to data.frame of model.full data (required for some model types).}

\item{anova}{Logical, whether to produce an Anova table with the RESI columns added. By default = `TRUE`.}

\item{summary}{Logical, whether to produce a summary table with the RESI columns added. By default = `TRUE`.}

\item{nboot}{Numeric, the number of bootstrap replicates. By default, 1000.}

\item{boot.method}{String, which type of bootstrap to use: `nonparam` = non-parametric bootstrap (default); `bayes` = Bayesian bootstrap.}

\item{vcovfunc}{The variance estimator function for constructing the Wald test statistic. By default, sandwich::vcovHC (the robust (sandwich) variance estimator).}

\item{alpha}{Numeric, significance level of the constructed CIs. By default, 0.05.}
}
\value{
Returns a list that includes function arguments, RESI point estimates, and confidence intervals in summary/anova-style tables
}
\description{
This function will estimate the robust effect size (RESI) from Vandekar, Rao, & Blume (2020) and its confidence interval in various ways for a fitted model object. The overall RESI is estimated via a Wald test. RESI is (optionally) estimated for each factor in summary-style table. RESI is (optionally) estimated for each variable/interaction in an Anova-style table for models with existing Anova methods. CIs can be calculated using either non-parametric or Bayesian bootstrapping.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: RESI point and interval estimation for models

\item \code{nls}: RESI point and interval estimation for nls models

\item \code{zeroinfl}: RESI point and interval estimation for hurdle models

\item \code{geeglm}: RESI point and interval estimation for GEE models

\item \code{lme}: RESI point and interval estimation for LME models
}}

