% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resi_pe.R
\name{resi_pe}
\alias{resi_pe}
\alias{resi_pe.default}
\alias{resi_pe.glm}
\alias{resi_pe.lm}
\alias{resi_pe.nls}
\alias{resi_pe.survreg}
\alias{resi_pe.hurdle}
\alias{resi_pe.coxph}
\alias{resi_pe.geeglm}
\alias{resi_pe.gee}
\alias{resi_pe.lme}
\title{Robust Effect Size index (RESI) point estimation}
\usage{
resi_pe(model.full, ...)

\method{resi_pe}{default}(
  model.full,
  model.reduced = NULL,
  data,
  summary = TRUE,
  vcovfunc = sandwich::vcovHC
)

\method{resi_pe}{glm}(
  model.full,
  model.reduced = NULL,
  data,
  anova = TRUE,
  summary = TRUE,
  vcovfunc = sandwich::vcovHC,
  ...
)

\method{resi_pe}{lm}(
  model.full,
  model.reduced = NULL,
  data,
  anova = TRUE,
  summary = TRUE,
  vcovfunc = sandwich::vcovHC,
  ...
)

\method{resi_pe}{nls}(
  model.full,
  model.reduced = NULL,
  data,
  summary = TRUE,
  vcovfunc = regtools::nlshc,
  ...
)

\method{resi_pe}{survreg}(
  model.full,
  model.reduced = NULL,
  data,
  anova = TRUE,
  summary = TRUE,
  vcovfunc = vcov,
  ...
)

\method{resi_pe}{hurdle}(
  model.full,
  model.reduced = NULL,
  data,
  summary = TRUE,
  vcovfunc = sandwich::sandwich,
  ...
)

\method{resi_pe}{coxph}(
  model.full,
  model.reduced = NULL,
  data,
  anova = TRUE,
  summary = TRUE,
  vcovfunc = vcov,
  ...
)

\method{resi_pe}{geeglm}(model.full, ...)

\method{resi_pe}{gee}(model.full, ...)

\method{resi_pe}{lme}(model.full, ...)
}
\arguments{
\item{model.full}{The lme model object}

\item{...}{Other arguments to be passed to Anova function}

\item{model.reduced}{the reduced model to compare with the full model. By default `NULL`, it's the same model as the full model but only having intercept.}

\item{data}{required data frame or object coercible to data frame of model.full data}

\item{summary}{whether to produce a summary table with the RESI columns added. By default = `TRUE`}

\item{vcovfunc}{the variance estimator function for constructing the Wald test statistic. By default, sandwich::vcovHC (the robust (sandwich) variance estimator)}

\item{anova}{whether to produce an Anova table with the RESI columns added. By default = `TRUE`}
}
\value{
Returns a list containing RESI point estimates

returns the summary-type table with RESI estimate for each factor
}
\description{
Robust Effect Size index (RESI) point estimation
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Robust Effect Size index (RESI) point estimation

\item \code{glm}: Robust Effect Size index (RESI) point estimation for generalized linear models

\item \code{lm}: Robust Effect Size index (RESI) point estimation for linear models

\item \code{nls}: Robust Effect Size index (RESI) point estimation for nonlinear least squares models

\item \code{survreg}: Robust Effect Size index (RESI) point estimation for survreg

\item \code{hurdle}: Robust Effect Size index (RESI) point estimation for hurdle models

\item \code{coxph}: Robust Effect Size index (RESI) point estimation for coxph models

\item \code{geeglm}: Robust Effect Size index (RESI) point estimation for geeglm object
This function calculates the RESI from geeglm model object

\item \code{gee}: Robust Effect Size index (RESI) point estimation for gee object
This function calculates the RESI from gee model object

\item \code{lme}: Robust Effect Size index (RESI) point estimation for lme object
This function calculate the RESI from lme model object
}}

