% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{S2fsq}
\alias{S2fsq}
\title{Covert S to Cohen's \emph{f}^2}
\usage{
S2fsq(S)
}
\arguments{
\item{S}{Numeric,the robust effect size index.}
}
\value{
Returns an estimate of Cohen's \emph{f}^2 based on the RESI
}
\description{
Converts robust effect size index (S) to Cohen's \emph{f}^2
(effect size for multiple regression) using the formula from Vandekar, Rao, & Blume (2020).
}
\keyword{power}
