% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S.R, R/internal_functions.R
\name{z2S}
\alias{z2S}
\title{Compute the robust effect size index estimator from Z statistic}
\usage{
z2S(z, n)

z2S(z, n)
}
\arguments{
\item{z}{The Z statistic for the parameter of interest.}

\item{n}{Number of independent samples.}
}
\value{
Returns a scalar or vector argument of the the robust effect size index estimate.

Returns a scalar or vector argument of the the robust effect size index estimate.
}
\description{
This function computes the robust effect size index from Vandekar, Rao, & Blume (2020).
Vector arguments are accepted. If different length arguments are passed they are dealt with in the usual way of R.

This function computes the robust effect size index from Vandekar, Rao, & Blume (2020).
Vector arguments are accepted. If different length arguments are passed they are dealt with in the usual way of R.
}
\keyword{power}
